checkpoint: BUG-006 git-mode passthrough; BUG-007 autocommit isolation; BUG-008 dry-run output

+++ crates/agt/src/scanner.rs
@@ -3,15 +3,17 @@ use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 use std::time::SystemTime;
 
-pub fn scan_modified_files(root: &Path, since_timestamp: i64) -> Result<Vec<PathBuf>> {
-    let threshold = SystemTime::UNIX_EPOCH + std::time::Duration::from_secs(since_timestamp as u64);
+pub fn scan_modified_files(root: &Path, since_timestamp: u64) -> Result<Vec<PathBuf>> {
+    let threshold = SystemTime::UNIX_EPOCH + std::time::Duration::from_secs(since_timestamp);
     let mut files = Vec::new();
 
     for entry in jwalk::WalkDir::new(root)
         .skip_hidden(false)
         .process_read_dir(|_depth, _path, _state, children| {
             children.retain(|entry| {
-                entry.as_ref().map_or(true, |dir_entry| dir_entry.file_name != OsStr::new(".git"))
+                entry
+                    .as_ref()
+                    .map_or(true, |dir_entry| dir_entry.file_name != OsStr::new(".git"))
             });
         })
         .into_iter()
